+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow users to join and leave an organization [#226](https://github.com/PublicMapping/districtbuilder/pull/578)
 - Require users to confirm before joining an organization [#593](https://github.com/PublicMapping/districtbuilder/pull/593)
 - Show not found page for missing projects and organizations [#613](https://github.com/PublicMapping/districtbuilder/pull/613)
+- Add paintbrush drawing tool [#611](https://github.com/PublicMapping/districtbuilder/pull/611)
 
 ### Changed